backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# WebSockets
websockets==12.0

# Additional utilities
python-dotenv==1.0.0
python-docx==1.1.0

# OpenAI API for agent execution
openai==1.3.7
PyPDF2==3.0.1
